+++ src/core/settings.py
@@ -48,6 +48,7 @@ INSTALLED_APPS = (
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "solo",
     "modeltranslation",
     "django_cotton",
     "base",