+++ src/home/admin.py
@@ -1,10 +1,11 @@
 from django.contrib.admin import ModelAdmin, register
+from solo.admin import SingletonModelAdmin
 
 from home.models import Experience, PersonalInfo
 
 
 @register(PersonalInfo)
-class PersonalInfoAdmin(ModelAdmin):
+class PersonalInfoAdmin(SingletonModelAdmin):
     pass