+++ src/home/models.py
@@ -6,10 +6,10 @@ from typing import NamedTuple
 
 from django.db import models
 from django.utils.translation import gettext, ngettext
+from solo.models import SingletonModel
 
 
-# TODO: make row unique
-class PersonalInfo(models.Model):  # type: ignore[django-manager-missing] # https://github.com/typeddjango/django-stubs/issues/1023
+class PersonalInfo(SingletonModel):  # type: ignore[django-manager-missing] # https://github.com/typeddjango/django-stubs/issues/1023
     name = models.CharField(max_length=100)
     description = models.CharField(max_length=100)
     biography = models.TextField()